refactor middleware to render board
