+++ client/components/Board.jsx
@@ -1,15 +1,33 @@
-import React, { useState } from 'react';
+import React, { useState, useEffect } from 'react';
 import { DragDropContext } from 'react-beautiful-dnd';
 import { v4 as uuidv4 } from 'uuid';
 import Category from './Category';
 import Users from './Users';
+import { api } from '../utils/api';
 
-const initialCategories = {
-  [uuidv4()]: {
-    name: 'Todo',
-    items: [],
-  },
-};
+const initialCategories = {};
+let initialUsers = [];
+async function getCategory(catUpdate) {
+  let categories = await api.getCategory();
+  initialUsers = await api.getUser();
+  const tasks = await api.getTask();
+
+  if (categories.length === 0) {
+    api.createCategory({ category: 'ToDo' });
+    categories = await api.getCategory();
+  }
+
+  for (let i = 0; i < categories.length; i++) {
+    const _id = categories[i]._id;
+    const name = categories[i].category;
+    const catTasks = tasks.filter((el) => el.Category === name);
+    catUpdate[_id] = {
+      name: name,
+      items: catTasks,
+    };
+  }
+}
+await getCategory(initialCategories);
 
 const onDragEnd = (result, categories, setCategories, users, setUsers) => {
   const { source, destination } = result;
@@ -28,6 +46,7 @@ const onDragEnd = (result, categories, setCategories, users, setUsers) => {
   } else if (source.droppableId === destination.droppableId) {
     // Reordering tasks within the same category
     const category = categories[source.droppableId];
+    console.log(category);
     console.log(categories);
     console.log(categories[source.droppableId]);
     const copiedItems = [...category.items];
@@ -67,22 +86,55 @@ const onDragEnd = (result, categories, setCategories, users, setUsers) => {
   }
 };
 
-const Board = () => {
+export default function Board() {
   const [categories, setCategories] = useState(initialCategories);
-  const [users, setUsers] = useState([]);
+  const [users, setUsers] = useState(initialUsers);
+  const [effect, setEffect] = useState([]);
+
+  useEffect(() => {
+    console.log('USE EFFECT IS BEING TRIGGEREDDDDDDD');
+    const newCats = Object.assign({}, categories);
+
+    async function updateCatTask() {
+      const tasks = await api.getTask();
+      const categories = await api.getCategory();
+      for (let i = 0; i < categories.length; i++) {
+        const _id = categories[i]._id;
+        const name = categories[i].category;
+        const catTasks = tasks.filter((el) => el.Category === name);
+        newCats[_id] = {
+          name: name,
+          items: catTasks,
+        };
+      }
+      setCategories(newCats);
+    }
+    updateCatTask();
+
+    setUsers(users);
+  }, [users, effect]);
 
   const addNewCategory = () => {
-    const newId = uuidv4();
+    const category = api.createCategory({ category: 'New Category' });
+    const { _id } = category;
+
     setCategories({
       ...categories,
-      [newId]: {
+      [_id]: {
         name: 'New Category',
         items: [],
       },
     });
   };
 
-  const addNewTask = (categoryId, task) => {
+  const addNewTask = async (categoryId, task) => {
+    if (!task) {
+      const { _id } = categoryId;
+      const category = categories[_id];
+      await api.createTask({ Task_Name: ' ', Category: category.name });
+
+      return;
+    }
     const category = categories[categoryId];
     const newItems = [...category.items, task];
     setCategories({
@@ -95,40 +147,30 @@ const Board = () => {
   };
 
   const addNewUser = (user) => {
-    // console.log('New user:', user);
-    // console.log('Adding to: ', users);
     setUsers((users) => {
       const updatedUsers = [...users, user];
       // console.log('New user list:', updatedUsers);
       return updatedUsers;
     });
   };
-  
-  const removeTask = (categoryId, removeTask) => {
-    const category = categories[categoryId];
-    const newItems = [];
 
-    setCategories({
-      ...categories,
-      [categoryId]: {
-        ...category,
-        items: newItems,
-      },
-    });
+  const removeTask = (categoryId, removeTask) => {
+    setEffect([]);
   };
 
-  const removeUser = (userId) => {
-    setUsers((prevUsers) => {
-      const updatedUsers = prevUsers.filter((user) => user._id !== userId);
-      console.log('Users before:', prevUsers);
-      console.log('Users after:', updatedUsers);
-      return updatedUsers;
-    });
+  const removeUser = async (userId) => {
+    console.log(userId);
+    await api.removeUser(userId);
+    setEffect([]);
   };
 
   const editTask = (categoryId, edittedTask) => {
     const category = categories[categoryId];
     const newItems = edittedTask;
+    console.log(
+      'The category id in the editTask definitition is App.js is: ',
+      categoryId
+    );
 
     setCategories({
       ...categories,
@@ -140,22 +182,40 @@ const Board = () => {
   };
 
   return (
-    <div className='board'>
+    <div className="app">
       <DragDropContext
-        onDragEnd={(result) => onDragEnd(result, categories, setCategories, users, setUsers)}
+        onDragEnd={(result) =>
+          onDragEnd(result, categories, setCategories, users, setUsers)
+        }
       >
-        <div className='categories-container'>
-          <Users userId={'usersCategory'} users={users} addNewUser={addNewUser} removeUser={removeUser} />
+        <div className="categories-container">
+          <Users
+            userId={'usersCategory'}
+            users={users}
+            addNewUser={addNewUser}
+            removeUser={removeUser}
+          />
           {Object.entries(categories).map(([id, category]) => (
-            <Category key={id} categoryId={id} category={category} addNewTask={addNewTask} removeTask={removeTask} editTask={editTask}/>
+            <Category
+              key={id}
+              categoryId={id}
+              category={category}
+              addNewTask={addNewTask}
+              removeTask={removeTask}
+              editTask={editTask}
+            />
           ))}
-          <div className='add-category-container'>
-            <button onClick={addNewCategory} className="add-category-button"> + New Section</button>
+          <div className="add-category-container">
+            <button
+              onClick={addNewCategory}
+              className="add-category-button"
+            >
+              {' '}
+              + New Section
+            </button>
           </div>
         </div>
       </DragDropContext>
     </div>
   );
 }
-
-export default Board;
