+++ package.json
@@ -15,6 +15,7 @@
   "license": "ISC",
   "dependencies": {
     "axios": "^1.5.0",
+    "bcryptjs": "^2.4.3",
     "cors": "^2.8.5",
     "eslint-config-airbnb": "^19.0.4",
     "express": "^4.18.2",
@@ -48,15 +49,3 @@
     "webpack-dev-server": "^4.15.1"
   }
 }
-
-
-
-
-
-
-
-
-
-
-
-
