+++ server/server.js
@@ -3,6 +3,7 @@ const mongoose = require('mongoose');
 const express = require('express');
 const path = require('path');
 const cors = require('cors');
+const bcrypt = require('bcryptjs');
 
 const PORT = 3000;
 const app = express();
